requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bert-score==0.3.13
BLEURT==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
datasets==2.13.1
dill==0.3.6
entmax==1.3
exceptiongroup==1.2.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
jsonargparse==3.13.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lightning-utilities==0.11.2
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
metricx @ git+https://github.com/sweta20/metricx.git@f53d7953b1feb57bb4eb04b33706f26146c113e7
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.14
namex==0.0.8
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
sacrebleu==2.4.2
safetensors==0.4.3
scipy==1.13.0
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tf-slim==1.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.3.0
torchmetrics==0.10.3
tqdm==4.66.2
transformers==4.30.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
unbabel-comet @ git+https://github.com/Unbabel/COMET.git@87866f29a80d7b3f1a9c68906ef47b4d23170f95
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
